[fix]Rename MuiThemeProvider -> ThemeProvider
