+++ core/src/main/kotlin/materialui/styles/themeprovider/themeProvider.kt
@@ -1,4 +1,4 @@
-package materialui.styles.muithemeprovider
+package materialui.styles.themeprovider
 
 import materialui.styles.muitheme.MuiTheme
 import react.RBuilder
@@ -12,5 +12,5 @@ private external val stylesModule: dynamic
 @Suppress("UnsafeCastFromDynamic")
 private val themeProviderComponent: RComponent<RProps, RState> = stylesModule.ThemeProvider
 
-fun RBuilder.muiThemeProvider(theme: MuiTheme, block: MuiThemeProviderBuilder.() -> Unit)
-    = child(MuiThemeProviderBuilder(themeProviderComponent, theme).apply(block).create())
+fun RBuilder.themeProvider(theme: MuiTheme, block: ThemeProviderBuilder.() -> Unit)
+    = child(ThemeProviderBuilder(themeProviderComponent, theme).apply(block).create())
