+++ sample/src/main/kotlin/demo/App.kt
@@ -8,9 +8,9 @@ import kotlinx.css.Color
 import materialui.styles.createMuiTheme
 import materialui.styles.muitheme.MuiTheme
 import materialui.styles.muitheme.options.palette
-import materialui.styles.muithemeprovider.muiThemeProvider
 import materialui.styles.palette.options.main
 import materialui.styles.palette.options.primary
+import materialui.styles.themeprovider.themeProvider
 import react.RBuilder
 import react.RComponent
 import react.RProps
@@ -19,7 +19,7 @@ import react.dom.*
 
 class App : RComponent<RProps, RState>() {
     override fun RBuilder.render() {
-        muiThemeProvider(theme) {
+        themeProvider(theme) {
             header {
                 Header.render(this)
             }
